fix: base structure
